+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 - Updated the `proto plugins` command to include the plugin's version when applicable.
 - WASM API
   - Added `install_global`, `uninstall_global`, and `native_uninstall` plugin functions.
+  - Added `pre_install`, `post_install`, `pre_run`, and `post_run` plugin hooks.
   - Added `plugin_version` field to `ToolMetadataOutput`.
+  - Added `InstallHook` and `RunHook` JSON structs.
 
 ## 0.14.1